task_1_semantic_search/semantic-search-twitter/requirements.txt
faiss-cpu
sentence-transformers
numpy
python-dotenv
tqdm
openai
